refactor(apollo error): created reusable hook to toast.error apollo error on application
